Introduce "auto-skipping" ability
Automatic character skipping in an ability to jump over groups of fixed symbols when hitting Backspace in the middle of the string.
This feature also allows trimming those fixed characters when editing the end of the string.

This commit also includes an important refactoring which moves the "autocomplete" flag into the CaretGravity.